+++ CHANGELOG.md
@@ -18,6 +18,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- webapp: Fix data mapping for initial state
 - database: Speedup id grouping
 - fetch: Fix query with polluted text cache
 - export: Fix query with polluted text cache